+++ test/gospn_test.go
@@ -1,14 +1,14 @@
 package test
 
 import (
-	"../pkg/parser"
-	"../pkg/petrinet"
+	"github.com/okamumu/gospn/pkg/parser"
+	"github.com/okamumu/gospn/pkg/petrinet"
 	"testing"
 )
 
 func BenchmarkGoSPNP1(b *testing.B) {
 	// for i := 0; i < b.N; i++ {
-	if net, imark, err := parser.PNreadFromFile("./data/spnp_example1.spn"); err == nil {
+	if net, imark, err := parser.PNreadFromFile("../example/spnp_example1.spn"); err == nil {
 		b.ResetTimer()
 		mg := petrinet.CreateMarkingGraphWithDFS(net, imark)
 		mg.TransMatrix()
@@ -21,7 +21,7 @@ func BenchmarkGoSPNP1(b *testing.B) {
 
 func BenchmarkGoSPNP2(b *testing.B) {
 	// for i := 0; i < b.N; i++ {
-	if net, imark, err := parser.PNreadFromFile("./data/spnp_example2.spn"); err == nil {
+	if net, imark, err := parser.PNreadFromFile("../example/spnp_example2.spn"); err == nil {
 		b.ResetTimer()
 		mg := petrinet.CreateMarkingGraphWithDFS(net, imark)
 		mg.TransMatrix()
@@ -34,7 +34,7 @@ func BenchmarkGoSPNP2(b *testing.B) {
 
 func BenchmarkGoSPNP3(b *testing.B) {
 	// for i := 0; i < b.N; i++ {
-	if net, imark, err := parser.PNreadFromFile("./data/spnp_example3.spn"); err == nil {
+	if net, imark, err := parser.PNreadFromFile("../example/spnp_example3.spn"); err == nil {
 		b.ResetTimer()
 		mg := petrinet.CreateMarkingGraphWithDFS(net, imark)
 		mg.TransMatrix()
@@ -47,7 +47,7 @@ func BenchmarkGoSPNP3(b *testing.B) {
 
 func BenchmarkGoSPNP4(b *testing.B) {
 	// for i := 0; i < b.N; i++ {
-	if net, imark, err := parser.PNreadFromFile("./data/spnp_example4.spn"); err == nil {
+	if net, imark, err := parser.PNreadFromFile("../example/spnp_example4.spn"); err == nil {
 		b.ResetTimer()
 		mg := petrinet.CreateMarkingGraphWithDFS(net, imark)
 		mg.TransMatrix()
@@ -60,7 +60,7 @@ func BenchmarkGoSPNP4(b *testing.B) {
 
 func BenchmarkGoSPNP5(b *testing.B) {
 	// for i := 0; i < b.N; i++ {
-	if net, imark, err := parser.PNreadFromFile("./data/spnp_example5.spn"); err == nil {
+	if net, imark, err := parser.PNreadFromFile(".../example/spnp_example5.spn"); err == nil {
 		b.ResetTimer()
 		mg := petrinet.CreateMarkingGraphWithDFS(net, imark)
 		mg.TransMatrix()
@@ -73,7 +73,7 @@ func BenchmarkGoSPNP5(b *testing.B) {
 
 func BenchmarkGoSPNP6(b *testing.B) {
 	// for i := 0; i < b.N; i++ {
-	if net, imark, err := parser.PNreadFromFile("./data/spnp_example6.spn"); err == nil {
+	if net, imark, err := parser.PNreadFromFile("../example/spnp_example6.spn"); err == nil {
 		b.ResetTimer()
 		mg := petrinet.CreateMarkingGraphWithDFS(net, imark)
 		mg.TransMatrix()
@@ -86,7 +86,7 @@ func BenchmarkGoSPNP6(b *testing.B) {
 
 func BenchmarkGoSPNP7(b *testing.B) {
 	// for i := 0; i < b.N; i++ {
-	if net, imark, err := parser.PNreadFromFile("./data/raid6.spn"); err == nil {
+	if net, imark, err := parser.PNreadFromFile("../example/raid6.spn"); err == nil {
 		b.ResetTimer()
 		mg := petrinet.CreateMarkingGraphWithDFS(net, imark)
 		mg.TransMatrix()
@@ -99,7 +99,7 @@ func BenchmarkGoSPNP7(b *testing.B) {
 
 func BenchmarkGoSPNP8(b *testing.B) {
 	// for i := 0; i < b.N; i++ {
-	if net, imark, err := parser.PNreadFromFile("./data/raid10.spn"); err == nil {
+	if net, imark, err := parser.PNreadFromFile("../example/raid10.spn"); err == nil {
 		b.ResetTimer()
 		mg := petrinet.CreateMarkingGraphWithDFS(net, imark)
 		mg.TransMatrix()