+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+# gospn 0.10.2
+
+- add a binary to convert MxGraph data to Petrinet definition file
+
+# gospn 0.10.1
+
+- build with go 1.16
+- use go.mod
+- provide Apple m1 binary
+
 # gospn 0.10.0
 
 - enhancement: generate block matrix